adding a name to the user model.
a bunch of active record associations that help:
- find start and end times for events.
- find v2_reservations for users and people
- find start and end times for reservations

all so that we can filter out reserved slots that we offer to people.

We don't want a user to be double booked: two reservations that overlap
nor do we want people to be double booked.

Currently, there are no model validations that prevent this.

using a specific version of elasticsearch in travis-ci

more attempting to get travis to work